refactor(controller): split gateway cleanup helpers
